Allow user configurable backup-enabled saving

+++ ValveLocalizationFile.cs
@@ -313,12 +313,16 @@ namespace VGUILocalizationTool
         public void WriteData(string local, List<LocalizationData> data)
         {
             string fileName = GetLocalFileName(local);
-            string fileNameBak = fileName + ".bak";
-            if (File.Exists(fileNameBak))
+            bool shouldSaveBak = global::VGUILocalizationTool.Properties.Settings.Default.SaveBackup;
+            if (shouldSaveBak)
             {
-                File.Delete(fileNameBak);
+                string fileNameBak = fileName + ".bak";
+                if (File.Exists(fileNameBak))
+                {
+                    File.Delete(fileNameBak);
+                }
+                File.Move(fileName, fileNameBak);
             }
-            File.Move(fileName, fileNameBak);
             //local = local.Substring(0, 1).ToUpper() + local.Substring(1);
             using (StreamWriter sw = new StreamWriter(fileName, false, System.Text.Encoding.Unicode))
             {
